refactor(dotnet-broker-kafka): add core function createMessageBrokerControllerFile
