+++ plugins/dotnet-broker-kafka/src/core/index.ts
@@ -3,3 +3,4 @@ export { createMessageBroker } from "./create-message-broker-method";
 export { getMessageBrokerName } from "./get-message-broker-name";
 export { createConsumerServiceFile } from "./create-consumer-service-file";
 export { createProducerServiceFile } from "./create-producer-service-file";
+export { createMessageBrokerControllerFile } from "./create-message-broker-controller-file";
